removed `lu-optins` component renamed luOptionsModule to luforoptions module
