+++ packages/ng/libraries/core/src/lib/api/select/input/api-select-input.component.html
@@ -7,7 +7,5 @@
 </div>
 <lu-option-picker>
 	<lu-api-paged-searcher></lu-api-paged-searcher>
-	<lu-options>
-		<lu-option *luDisplayer="let option" [value]="option">{{option.name}}</lu-option>
-	</lu-options>
+	<lu-option *luForOptions="let option" [value]="option">{{option.name}}</lu-option>
 </lu-option-picker>
