+++ packages/ng/libraries/core/src/lib/option/operator/for-options/for-options.directive.ts
