+++ packages/ng/libraries/core/src/lib/option/operator/for-options/for-options.module.ts
@@ -1,19 +1,14 @@
 import { NgModule } from '@angular/core';
-import { LuOptionsComponent } from './options.component';
-import { CommonModule } from '@angular/common';
 import { LuForOptionsDirective } from './for-options.directive';
 
 @NgModule({
 	imports: [
-		CommonModule,
 	],
 	declarations: [
-		LuOptionsComponent,
 		LuForOptionsDirective,
 	],
 	exports: [
-		LuOptionsComponent,
 		LuForOptionsDirective,
 	],
 })
-export class LuOptionsModule {}
+export class LuForOptionsModule {}
