+++ packages/ng/libraries/core/src/lib/option/operator/for-options/index.ts
@@ -1,0 +1,2 @@
+export * from './for-options.directive';
+export * from './for-options.module';
