+++ packages/ng/libraries/core/src/lib/option/operator/index.ts
@@ -1,6 +1,6 @@
 export * from './feeder/index';
 export * from './pager/index';
 export * from './searcher/index';
-export * from './options/index';
+export * from './for-options/index';
 export * from './option-operator.module';
 export * from './option-operator.model';
