--- packages/ng/libraries/core/src/lib/option/operator/options/index.ts
@@ -1,3 +1,0 @@
-export * from './for-options.directive';
-export * from './options.component';
-export * from './options.module';
