--- packages/ng/libraries/core/src/lib/option/operator/options/options.component.html
@@ -1,3 +1,0 @@
-<ng-template ngFor [ngForOf]="options$ | async" let-option [ngForTrackBy]="trackByFn">
-	<ng-container *ngTemplateOutlet="displayTemplate; context: {$implicit: option}"></ng-container>
-</ng-template>
