--- packages/ng/libraries/core/src/lib/option/operator/options/options.component.scss
@@ -1,5 +1,0 @@
-:host {
-	display: block;
-	overflow-y: auto;
-	flex: 1 1 auto;
-}
