+++ packages/ng/libraries/core/src/lib/user/select/input/user-select-input.component.html
@@ -7,7 +7,5 @@
 </div>
 <lu-option-picker>
 	<lu-user-paged-searcher></lu-user-paged-searcher>
-	<lu-options>
-		<lu-option *luDisplayer="let user" [value]="user">{{ user | luUserDisplay: searchFormat }}</lu-option>
-	</lu-options>
+	<lu-option *luForOptions="let user" [value]="user">{{ user | luUserDisplay: searchFormat }}</lu-option>
 </lu-option-picker>
